@@ -668,7 +668,9 @@ class Data:
 
         ## check for zeros in model error
         for comp in ["zxx", "zxy", "zyx", "zyy", "tzx", "tzy"]:
-            find_zeros = np.where(self.dataframe[f"{comp}_model_error"] == 0)[0]
+            find_zeros = np.where(self.dataframe[f"{comp}_model_error"] == 0)[
+                0
+            ]
             if find_zeros.shape[0] > 0:
                 if comp in ["zxx", "zxy", "zyx", "zyy"]:
                     error_percent = self.z_model_error.error_value
@@ -859,7 +861,8 @@ class Data:
                         self.wave_sign_tipper = hline[hline.find("(") + 1]
 
                 elif (
-                    len(hline[1:].strip().split()) >= 2 and hline.count(".") > 0
+                    len(hline[1:].strip().split()) >= 2
+                    and hline.count(".") > 0
                 ):
                     value_list = [
                         float(value) for value in hline[1:].strip().split()
@@ -983,7 +986,7 @@ class Data:
             if len(inv_mode_list) != inv_list:
                 continue
             else:
-                tf_arr = np.zeros(len(inv_list), dtype=np.bool)
+                tf_arr = np.zeros(len(inv_list), dtype=bool)
 
                 for tf, data_inv in enumerate(inv_list):
                     if data_inv in self.inv_mode_dict[inv_key]:
@@ -1140,7 +1143,9 @@ class Data:
             lines = fid.readlines()
 
         def fix_line(line_list):
-            return " ".join("".join(line_list).replace("\n", "").split()) + "\n"
+            return (
+                " ".join("".join(line_list).replace("\n", "").split()) + "\n"
+            )
 
         h1 = fix_line(lines[0:n])
         h2 = fix_line(lines[n : 2 * n])
