@@ -67,9 +67,9 @@ author = "Jared Peacock, Alison Kirkby, Fei Zhang, Lars Krieger, Stephan Thiel"
 #
 
 # The short X.Y version.
-version = "2.0.3"
+version = "2.0.5"
 # The full version, including alpha/beta/rc tags.
-release = "2.0.3"
+release = "2.0.5"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
