@@ -12,16 +12,6 @@ from .convariance import Covariance
 
 from .config import ModEMConfig
 
-# from .model_manipulator import ModelManipulator
-# from .plot_response import PlotResponse
-
-# # from .plot_pt_maps import PlotPTMaps
-# # from .plot_depth_slice import PlotDepthSlice
-# # from mtpy.imaging.modem_phase_tensor_maps import PlotPTMaps  # can cause circular import error
-# # from mtpy.imaging.plot_depth_slice import PlotDepthSlice     # can cause circular import error
-# # from .plot_slices import PlotSlices
-# # from .plot_rms_maps import PlotRMSMaps
-# # from .phase_tensor_maps import PlotPTMaps
 
 __all__ = [
     "ModEMError",
@@ -33,9 +23,5 @@ __all__ = [
     "ControlFwd",
     "Covariance",
     "ModEMConfig",
-    #     "ModelManipulator",
-    #     # "PlotResponse",
-    #     # "PlotSlices",
-    #     # "PlotRMSMaps"
-    #     # ,'PlotPTMaps', 'PlotDepthSlice'
+
 ]
